refactor(websocket): move WebSocket code from core to react package
- Move WebSocketManager, OptimisticEngine from @monorise/core to @monorise/react
- This separates frontend client code from backend Lambda code
- @monorise/core is now 100% backend-specific
- Add nanoid dependency to @monorise/react
- Export WebSocketManager and types from @monorise/react
- Fix type casting issues with Entity types
- Wrap state updates with immer produce

BREAKING CHANGE: WebSocketManager must now be imported from @monorise/react
