+++ packages/react/package.json
@@ -25,7 +25,8 @@
   "dependencies": {
     "@tabler/icons-react": "^3.30.0",
     "axios": "^1.8.1",
-    "immer": "^10.1.1"
+    "immer": "^10.1.1",
+    "nanoid": "^5.1.3"
   },
   "devDependencies": {
     "@types/react": "^19",
