+++ packages/react/websocket/WebSocketManager.ts
